Extract datasource query processing
