+++ src/datasource/index.ts
@@ -7,7 +7,6 @@ import {
   AdHocVariableFilter,
   CoreApp,
   DataFrame,
-  DataLink,
   DataQueryRequest,
   DataQueryResponse,
   DataSourceApi,
@@ -30,12 +29,11 @@ import {
   SupplementaryQueryType,
   TimeRange,
 } from '@grafana/data';
-import { BucketAggregation, DataLinkConfig, ElasticsearchQuery, TermsQuery, FieldCapabilitiesResponse } from './types';
+import { BucketAggregation, DataLinkConfig, ElasticsearchQuery, TermsQuery, FieldCapabilitiesResponse } from '@/types';
 import { 
   DataSourceWithBackend, 
   getTemplateSrv, 
-  TemplateSrv,
-  getDataSourceSrv } from '@grafana/runtime';
+  TemplateSrv } from '@grafana/runtime';
 import { QuickwitOptions } from 'quickwit';
 import { getDataQuery } from 'QueryBuilder/elastic';
 import { colors } from '@grafana/ui';
@@ -49,7 +47,8 @@ import ElasticsearchLanguageProvider from 'LanguageProvider';
 import { ReactNode } from 'react';
 import { fieldTypeMap } from 'utils';
 import { addAddHocFilter } from 'modifyQuery';
-import { LogContextProvider, LogRowContextOptions } from './LogContext/LogContextProvider';
+import { LogContextProvider, LogRowContextOptions } from '@/LogContext/LogContextProvider';
+import { getQueryResponseProcessor } from 'datasource/processResponse';
 
 export const REF_ID_STARTER_LOG_VOLUME = 'log-volume-';
 
@@ -94,16 +93,8 @@ export class QuickwitDataSource
   }
 
   query(request: DataQueryRequest<ElasticsearchQuery>): Observable<DataQueryResponse> {
-     return super.query(request)
-       .pipe(map((response) => {
-          response.data.forEach((dataFrame) => {
-            const metrics = request.targets[0]!.metrics
-            if (metrics && metrics[0].type === 'logs'){
-              enhanceDataFrameWithDataLinks(dataFrame, this.dataLinks, this.logMessageField);
-            }
-          });
-         return response;
-       }));
+    const queryProcessor = getQueryResponseProcessor(this, request)
+     return super.query(request) .pipe(map(queryProcessor.processResponse));
   }
 
     /**
@@ -739,81 +730,4 @@ function luceneEscape(value: string) {
   }
 
   return value.replace(/([\!\*\+\-\=<>\s\&\|\(\)\[\]\{\}\^\~\?\:\\/"])/g, '\\$1');
-}
-
-export function enhanceDataFrameWithDataLinks(dataFrame: DataFrame, dataLinks: DataLinkConfig[], logMessageField: string | undefined) {
-  // Ignore log volume dataframe, no need to add links or a displayed message field.
-  if (!dataFrame.refId || dataFrame.refId.startsWith('log-volume')) {
-    return;
-  }
-  if (logMessageField) {
-    const messageFields = logMessageField.split(',');
-    let field_idx_list = [];
-    for (const messageField of messageFields) {
-      const field_idx = dataFrame.fields.findIndex((field) => field.name === messageField);
-      if (field_idx !== -1) {
-        field_idx_list.push(field_idx);
-      }
-    }
-    const displayedMessages = Array(dataFrame.length);
-    for (let idx = 0; idx < dataFrame.length; idx++) {
-      let displayedMessage = "";
-      // If we have only one field, we assume the field name is obvious for the user and we don't need to show it.
-      if (field_idx_list.length === 1) {
-        displayedMessage = `${dataFrame.fields[field_idx_list[0]].values[idx]}`;
-      } else {
-        for (const field_idx of field_idx_list) {
-          displayedMessage += ` ${dataFrame.fields[field_idx].name}=${dataFrame.fields[field_idx].values[idx]}`;
-        }
-      }
-      displayedMessages[idx] = displayedMessage.trim();
-    }
-
-    const newField = {
-        name: 'message',
-        type: FieldType.string,
-        config: {},
-        values: displayedMessages,
-    }
-    const [timestamp, ...rest] = dataFrame.fields;
-    dataFrame.fields = [timestamp, newField, ...rest];
-  }
-  
-  if (!dataLinks.length) {
-    return;
-  }
-
-  for (const field of dataFrame.fields) {
-    const linksToApply = dataLinks.filter((dataLink) => dataLink.field === field.name);
-
-    if (linksToApply.length === 0) {
-      continue;
-    }
-
-    field.config = field.config || {};
-    field.config.links = [...(field.config.links || [], linksToApply.map(generateDataLink))];
-  }
-}
-
-function generateDataLink(linkConfig: DataLinkConfig): DataLink {
-  const dataSourceSrv = getDataSourceSrv();
-
-  if (linkConfig.datasourceUid) {
-    const dsSettings = dataSourceSrv.getInstanceSettings(linkConfig.datasourceUid);
-
-    return {
-      title: linkConfig.urlDisplayLabel || '',
-      url: '',
-      internal: {
-        query: { query: linkConfig.url },
-        datasourceUid: linkConfig.datasourceUid,
-        datasourceName: dsSettings?.name ?? 'Data source not found',
-      },
-    };
-  } else {
-    return {
-      title: linkConfig.urlDisplayLabel || '',
-      url: linkConfig.url,
-    };
-  }
 }
